backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
websockets==16.0
google-genai==1.70.0
faiss-cpu==1.13.2
numpy>=2.2.0
pdfplumber==0.11.9
pdf2image==1.17.0
Pillow>=11.0.0
weasyprint==68.1
Jinja2==3.1.6
python-dotenv==1.1.0
python-multipart==0.0.20
pydantic>=2.0.0
